requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
invoke==2.2.0
multidict==6.6.3
paramiko==4.0.0
pillow==11.3.0
propcache==0.3.2
pycparser==2.23
PyNaCl==1.6.0
requests==2.32.5
ts3API==0.9.4
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
